% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Find.L2ErrorMean}
\alias{Find.L2ErrorMean}
\title{Find.L2ErrorMean}
\usage{
Find.L2ErrorMean(YHat.Labels, Y.Lables, BinaryClassification)
}
\arguments{
\item{YHat.Labels, }{is the classification vector [I,n] you wish to test, where I is the number of Test causes, and n is the number of classifications}

\item{Y.Lables, }{is the true classification vector [n] you wish to test, where n is the number of classifications}

\item{BinaryClassification}{an integer value determining if the Data set is a BinaryClassification, with only 0 and 1's as the Labels for each data set}
}
\value{
The Mean error of each the Column [n], which is an average of each
}
\description{
Find.L2ErrorMean, helps you find the L2 mean of your Y.Hat Results from a MachineLearning Algorithm.
}
\examples{

}
